+++ apollo-router/src/plugins/mod.rs
@@ -30,6 +30,7 @@ mod expose_query_plan;
 mod forbid_mutations;
 mod headers;
 mod include_subgraph_errors;
+pub(crate) mod limits;
 pub(crate) mod override_url;
 pub(crate) mod rhai;
 pub(crate) mod subscription;
